Do not listen to touch events until client is initialized
Fixes #2148
